rename newapp
